Desktop/ac-back/requirements.txt
cryptography==3.0
wheel
Django==3.1
djangorestframework==3.11.1
markdown==3.2.2
django-filter==2.3.0
psycopg2==2.8.5
docxtpl==0.10.0
django-cors-headers==3.5.0
django-pgcrypto-fields==2.5.2
Pillow
pyOpenSSL==19.1.0
requests==2.24.0
celery==4.4.6
flower==0.9.5
python-dateutil==2.8.1
redis==3.5.3
sentry-sdk==0.17.0
python-u2flib-server==5.0.0
translitua==1.3.1
python-Levenshtein==0.12.0
django-taggit==1.3.0
openpyxl==3.0.5
drf-yasg==1.17.1
uwsgi==2.0.19.1
pypng==0.0.20
PyQRCode==1.2.1
twilio==6.45.0
xlrd==1.2.0
python-dotenv==0.14.0
boto3==1.14.48
django-redis==4.12.1
loguru==0.5.1
django-ses==1.0.2
django-prometheus==2.1.0
djangorestframework-simplejwt==4.4.0
django-clearcache==1.1.1
drf-nested-routers==0.91
workdays==1.4
turbosmsua==0.3
-e git://github.com/vhsporno/django-gm2m@develop#egg=django-gm2m
python-logstash
elastic-apm==5.8.1
django-cacheops==5.0.1
pyTelegramBotAPI